# parse_html.activity: about 30% speedup for html parsing