I've been reviewing some older takeouts, so had cachew off and parsing was a bit painful... so had a quick look in a profiler and did some optimization

A few optimizations
- using `.find` instead of `.select` is faster since it's not using CSS selectors
- using `SoupStrainer` is faster since it only does partial parsing and avoids materializing parts of soup we don't actually use

Measurements on a big `Chrome/MyActivity.html` file

- before
  - parsing (up to for loop over `outer_divs`: 17s
  - processing (everything in for loop): 16s

- after
  - parsing: 13s
  - procesing: 11s