@seanbreckenridge 2024-09-10
thanks!

will parse against my old takeouts just to make sure nothing breaks and merge in a bit